Stop using Type Script
